Add support for timelines with GPU compisiting turned ON

+++ app/helpers/timelines_helper.rb
@@ -11,7 +11,6 @@ module TimelinesHelper
 
     @testing_sections_list = get_testing_sections(timeline_events_data)
 
-
     # Return a hash of the performance metrics
     return aggregate_performance_metrics.to_json
 
@@ -82,17 +81,17 @@ module TimelinesHelper
   def insert_performanceMetrics(results_hash)
     metrics_hash = {}
 
-    # Get Frame rendering metrics
-    metrics_hash["renderingMetrics"] = get_rendering_metrics
+    # Get styling metrics
+    metrics_hash["calculateStylesMetrics"] = get_styling_metrics
 
     # Get Layout metrics
     metrics_hash["layoutMetrics"] = get_layout_metrics
 
+    # Get Frame rendering metrics
+    metrics_hash["renderingMetrics"] = get_rendering_metrics
+
     # Get painting metrics
     metrics_hash["paintMetrics"] = get_painting_metrics
-
-    # Get styling metrics
-    metrics_hash["calculateStylesMetrics"] = get_styling_metrics
 
     #Insert the results in the hash
     results_hash["performanceMetrics"] = metrics_hash
@@ -112,29 +111,40 @@ module TimelinesHelper
         # Check the events of an object in the section
         current_object = section[i]
 
-        # The true events list is the value of the "children" key
+        # The true events list is the value of the "children" key, if any
         events_list = current_object["children"]
 
-        # Check for the beginFrame event; when found:
-        # 1. push the timing of the object in the list of timings,
-        # 2. set the frame_event_pending flag
-        is_frame_event_found = check_for_event_type(events_list, "BeginFrame")
-
-        if is_frame_event_found.nil?
+        if events_list.nil?
           i += 1
         else
-          rendering_timings.append(current_object["endTime"] - current_object["startTime"])
-          is_frame_event_pending = false
+          # Check for the beginFrame event; when found:
+          # 1. push the timing of the object in the list of timings,
+          # 2. set the frame_event_pending flag
+          is_frame_event_found = check_for_event_type(events_list, "BeginFrame")
+
+          if is_frame_event_found.nil?
+            i += 1
+          else
+            rendering_timings.append(current_object["endTime"] - current_object["startTime"])
+            is_frame_event_pending = false
+          end
         end
       end
     end
 
     # create the hash of metrics
     rendering_metrics_hash = {}
-    rendering_metrics_hash["averageFrameRenderingTime"] = rendering_timings.sum.to_f / rendering_timings.length
-    rendering_metrics_hash["maxFrameRenderingTime"] = rendering_timings.max
-    rendering_metrics_hash["minFrameRenderingTime"] = rendering_timings.min
-    rendering_metrics_hash["frameRenderingRepetitions"] = rendering_timings.length
+    if rendering_timings.empty?
+      rendering_metrics_hash["averageFrameRenderingTime"] = 0
+      rendering_metrics_hash["maxFrameRenderingTime"] = 0
+      rendering_metrics_hash["minFrameRenderingTime"] = 0
+      rendering_metrics_hash["frameRenderingRepetitions"] = 0
+    else
+      rendering_metrics_hash["averageFrameRenderingTime"] = rendering_timings.sum.to_f / rendering_timings.length
+      rendering_metrics_hash["maxFrameRenderingTime"] = rendering_timings.max
+      rendering_metrics_hash["minFrameRenderingTime"] = rendering_timings.min
+      rendering_metrics_hash["frameRenderingRepetitions"] = rendering_timings.length
+    end
 
     return rendering_metrics_hash
   end
@@ -154,29 +164,40 @@ module TimelinesHelper
         # Check the events of an object in the section
         current_object = section[i]
 
-        # The true events list is the value of the "children" key
+        # The true events list is the value of the "children" key, if any
         events_list = current_object["children"]
 
-        # Check for the Layout event; when found:
-        # 1. push the timing of the object in the list of timings,
-        # 2. set the layout_event_pending flag
-        layout_event_found = check_for_event_type(events_list, "Layout")
-
-        if layout_event_found.nil?
+        if events_list.nil?
           i += 1
         else
-          layout_timings.append(layout_event_found["endTime"] - layout_event_found["startTime"])
-          is_layout_event_pending = false
+          # Check for the Layout event; when found:
+          # 1. push the timing of the object in the list of timings,
+          # 2. set the layout_event_pending flag
+          layout_event_found = check_for_event_type(events_list, "Layout")
+
+          if layout_event_found.nil?
+            i += 1
+          else
+            layout_timings.append(layout_event_found["endTime"] - layout_event_found["startTime"])
+            is_layout_event_pending = false
+          end
         end
       end
     end
 
     # create the hash of metrics
     layout_metrics_hash = {}
-    layout_metrics_hash["averageLayoutTime"] = layout_timings.sum.to_f / layout_timings.length
-    layout_metrics_hash["maxLayoutTime"] = layout_timings.max
-    layout_metrics_hash["minLayoutTime"] = layout_timings.min
-    layout_metrics_hash["layoutRepetition"] = layout_timings.length
+    if layout_timings.empty?
+      layout_metrics_hash["averageLayoutTime"] = 0
+      layout_metrics_hash["maxLayoutTime"] = 0
+      layout_metrics_hash["minLayoutTime"] = 0
+      layout_metrics_hash["layoutRepetition"] = 0
+    else
+      layout_metrics_hash["averageLayoutTime"] = layout_timings.sum.to_f / layout_timings.length
+      layout_metrics_hash["maxLayoutTime"] = layout_timings.max
+      layout_metrics_hash["minLayoutTime"] = layout_timings.min
+      layout_metrics_hash["layoutRepetition"] = layout_timings.length
+    end
 
     return layout_metrics_hash
   end
@@ -196,29 +217,40 @@ module TimelinesHelper
         # Check the events of an object in the section
         current_object = section[i]
 
-        # The true events list is the value of the "children" key
+        # The true events list is the value of the "children" key, if any
         events_list = current_object["children"]
 
-        # Check for the RecalculateStyles event; when found:
-        # 1. push the timing of the object in the list of timings,
-        # 2. set the style_event_pending flag
-        style_event_found = check_for_event_type(events_list, "RecalculateStyles")
-
-        if style_event_found.nil?
+        if events_list.nil?
           i += 1
         else
-          styling_timings.append(style_event_found["endTime"] - style_event_found["startTime"])
-          is_style_event_pending = false
+          # Check for the RecalculateStyles event; when found:
+          # 1. push the timing of the object in the list of timings,
+          # 2. set the style_event_pending flag
+          style_event_found = check_for_event_type(events_list, "RecalculateStyles")
+
+          if style_event_found.nil?
+            i += 1
+          else
+            styling_timings.append(style_event_found["endTime"] - style_event_found["startTime"])
+            is_style_event_pending = false
+          end
         end
       end
     end
 
     # create the hash of metrics
     styling_metrics_hash = {}
-    styling_metrics_hash["averageCalcStylesTime"] = styling_timings.sum.to_f / styling_timings.length
-    styling_metrics_hash["maxCalcStylesTime"] = styling_timings.max
-    styling_metrics_hash["minCalcStylesTime"] = styling_timings.min
-    styling_metrics_hash["calcStylesRepetitions"] = styling_timings.length
+    if styling_timings.empty?
+      styling_metrics_hash["averageCalcStylesTime"] = 0
+      styling_metrics_hash["maxCalcStylesTime"] = 0
+      styling_metrics_hash["minCalcStylesTime"] = 0
+      styling_metrics_hash["calcStylesRepetitions"] = 0
+    else
+      styling_metrics_hash["averageCalcStylesTime"] = styling_timings.sum.to_f / styling_timings.length
+      styling_metrics_hash["maxCalcStylesTime"] = styling_timings.max
+      styling_metrics_hash["minCalcStylesTime"] = styling_timings.min
+      styling_metrics_hash["calcStylesRepetitions"] = styling_timings.length
+    end
 
     return styling_metrics_hash
   end
@@ -241,38 +273,49 @@ module TimelinesHelper
         # Check the events of an object in the section
         current_object = section[i]
 
-        # The true events list is the value of the "children" key
+        # The true events list is the value of the "children" key, if any
         events_list = current_object["children"]
 
-        # Check for a paint events; when one is found:
-        # 1. push the timing of the object in the list of section painting timings
-        # 2. iterate over the other children of the object to find any other paint events
-        # 3. set the paint_events_pending flag
-        # 4. Add the section timings to the timings array
-        paint_event_found = check_for_event_type(events_list, "Paint")
-
-        if paint_event_found.nil?
+        if events_list.nil?
           i += 1
         else
-          section_painting_timings.append(paint_event_found["endTime"] - paint_event_found["startTime"])
+          # Check for a paint events; when one is found:
+          # 1. push the timing of the object in the list of section painting timings
+          # 2. iterate over the other children of the object to find any other paint events
+          # 3. set the paint_events_pending flag
+          # 4. Add the section timings to the timings array
+          paint_event_found = check_for_event_type(events_list, "Paint")
 
-          while !(another_paint_event = check_for_event_type(events_list, "Paint")).nil?
-            section_painting_timings.append(another_paint_event["endTime"] - another_paint_event["startTime"])
+          if paint_event_found.nil?
+            i += 1
+          else
+            section_painting_timings.append(paint_event_found["endTime"] - paint_event_found["startTime"])
+
+            while !(another_paint_event = check_for_event_type(events_list, "Paint")).nil?
+              section_painting_timings.append(another_paint_event["endTime"] - another_paint_event["startTime"])
+            end
+
+            paint_events_pending = false
+            painting_timings.append(section_painting_timings)
           end
-
-          paint_events_pending = false
-          painting_timings.append(section_painting_timings)
         end
       end
     end
 
     # create the hash of metrics
-    painting_timings = painting_timings.flatten
     painting_metrics_hash = {}
-    painting_metrics_hash["averagePaintTime"] = painting_timings.sum.to_f / painting_timings.length
-    painting_metrics_hash["maxPaintTime"] = painting_timings.max
-    painting_metrics_hash["minPaintTime"] = painting_timings.min
-    painting_metrics_hash["paintOccurrences"] = painting_timings.length
+    if painting_timings.empty?
+      painting_metrics_hash["averagePaintTime"] = 0
+      painting_metrics_hash["maxPaintTime"] = 0
+      painting_metrics_hash["minPaintTime"] = 0
+      painting_metrics_hash["paintOccurrences"] = 0
+    else
+      painting_timings = painting_timings.flatten
+      painting_metrics_hash["averagePaintTime"] = painting_timings.sum.to_f / painting_timings.length
+      painting_metrics_hash["maxPaintTime"] = painting_timings.max
+      painting_metrics_hash["minPaintTime"] = painting_timings.min
+      painting_metrics_hash["paintOccurrences"] = painting_timings.length
+    end
 
     return painting_metrics_hash
   end
@@ -287,8 +330,11 @@ module TimelinesHelper
       obj = list_of_objects[i]
       if obj.has_value?(event_name)
         event_found = obj
-        # for multiple paints remove the object from this list
-        list_of_objects.delete(obj)
+
+        if event_name == "Paint"
+          # for multiple paints remove the object from this list
+          list_of_objects.delete(obj)
+        end
       else
         i += 1
       end
